fix bugs with converting in XML format, create structure of server modules
