+++ server/developers/developers.json
@@ -1,1 +1,1 @@
-[{"name":"Tkachenko Polina","age":"19","lovely_color":"pink","exp":"nothing","hobbie":"Sewing, table games, stand up comedy","avatar":"img/1.jpg"},{"name":"Bondarenko Vyacheslav","age":"36","lovely_color":"turquoise","exp":"nothing","hobbie":"Design, it, sport","avatar":"img/1.jpg"},{"name":"Vakulenko Artem","age":"22","lovely_color":"blue","exp":"nothing","hobbie":"Music, literature","avatar":"img/tema.jpg"},{"name":"Skrypnik Maksym","age":"25","lovely_color":"blue","exp":"nothing","hobbie":"Hookah, football ","avatar":"img/max.jpg"},{"name":"Kalyna Olha","age":"18","lovely_color":"purple","exp":"nothing","hobbie":"Board games, podcasts, books","avatar":"img/1.jpg"},{"name":"Marharyta Horiunova","age":"31","lovely_color":"green","exp":"nothing","hobbie":"She's not with us, but she's part of the team","avatar":"img/1.jpg"}]
+[{"name":"Tkachenko Polina","age":"19","lovely_color":"pink","exp":"nothing","hobbie":"Sewing, table games, stand up comedy","avatar":"img/1.jpg"},{"name":"Bondarenko Vyacheslav","age":"36","lovely_color":"turquoise","exp":"nothing","hobbie":"Design, it, sport","avatar":"img/1.jpg"},{"name":"Vakulenko Artem","age":"22","lovely_color":"blue","exp":"nothing","hobbie":"Music, literature","avatar":"img/tema.jpg"},{"name":"Skrypnik Maksym","age":"25","lovely_color":"red","exp":"nothing","hobbie":"Hookah, football","avatar":"img/max.jpg"},{"name":"Kalyna Olha","age":"18","lovely_color":"purple","exp":"nothing","hobbie":"Board games, podcasts, books","avatar":"img/1.jpg"},{"name":"Marharyta Horiunova","age":"31","lovely_color":"green","exp":"nothing","hobbie":"She's not with us, but she's part of the team","avatar":"img/1.jpg"}]
