+++ server/questions/questions.xml
@@ -1,1 +1,1 @@
-<questions><question><questionText>123</questionText><theme>HTML</theme><date>1617980402897</date><stringDate>09.04.2021 | 18:00:02</stringDate><type>JSON,XML,YAML,CSV</type><answer>true</answer></question></questions>
+<questions></questions>
