+++ server/server.js
@@ -1,10 +1,5 @@
 var http = require("http");
-var fs = require("fs");
-var { json } = require("body-parser");
-var XMLparser = require("fast-xml-parser");
-var jsonParser = json();
-var yamlParser = require("js-yaml");
-var YAML = require("json-to-pretty-yaml");
+var {watchMethodAndUrl} = require("./server_modules/watch");
 
 var server = http.createServer(function (req, res) {
   var headers = {
@@ -13,328 +8,13 @@ var server = http.createServer(function (req, res) {
     "Access-Control-Allow-Methods": "GET,PUT,POST,DELETE",
     "Access-Control-Max-Age": 12344345789,
   };
-
-  if (req.method === "OPTIONS") {
-    res.writeHead(204, headers);
-    res.end();
-  }
-  if (req.url === "/developers" && req.method === "GET") {
-    var serverAnswer = fs.readFileSync("developers/developers.json", "utf-8");
-    res.writeHead(200, headers);
-    res.end(serverAnswer);
-  }
-  if (req.url === "/developers" && req.method === "PUT") {
-    jsonParser(req, res, (err) => {
-    var devBuff = fs.readFileSync('developers/developers.json' , 'utf-8');
-    var devArray = JSON.parse(devBuff);
-      for(var i = 0; i < devArray.length; i++){
-        if(req.body.name === devArray[i].name){
-          devArray[i].name = req.body.name;
-          devArray[i].age = req.body.age;
-          devArray[i].lovely_color = req.body.lovely_color;
-          devArray[i].exp = req.body.exp;
-          devArray[i].hobbie = req.body.hobbie;
-          devArray[i].avatar = req.body.avatar;
-        }
-      }
-      fs.writeFileSync('developers/developers.json', JSON.stringify(devArray));
-      res.writeHead(200, headers);
-      res.end(JSON.stringify(devArray));
-    });
-  }
-
-  if (req.url !== "/developers") {
     watchMethodAndUrl(req, res, headers);
-  }
+
 });
 
 server.listen(3000);
 
-function watchMethodAndUrl(req, res, headers) {
-  /* функция распределяющая распределяющая что делать, 
-  при определенное варианте запроса */
-  if (req.method === "GET") {
-    watchGetUrl(req, res, headers);
-  } else if (req.method === "POST") {
-    watchPostUrl(req, res, headers);
-  } else if (req.method === "DELETE") {
-    watchDeleteUrl(req, res, headers);
-  } else if (req.method === "OPTIONS") {
-    res.writeHead(204, headers);
-    res.end();
-    return;
-  }
-}
-function watchGetUrl(req, res, headers) {
-  /* функция отвечающая за обработку get запросов, 
-  смотря на query параметры, которые пришли в запросе */
-  var URL = new URLSearchParams(req.url);
-  if (URL.get("type") === "JSON") {
-    getFromXMLFile(URL);
-    var serverAnswer = getFromJSONFile(URL);
-    res.writeHead(200, headers);
-    res.end(JSON.stringify(serverAnswer));
-  }
-  if (URL.get("type") === "XML") {
-    var serverAnswer = getFromXMLFile(URL);
-    if (serverAnswer === undefined) {
-      res.writeHead(200, headers);
-      res.end(JSON.stringify([]));
-    } else if (!Array.isArray(serverAnswer)) {
-      res.writeHead(200, headers);
-      res.end(JSON.stringify([serverAnswer]));
-    } else {
-      res.writeHead(200, headers);
-      res.end(JSON.stringify(serverAnswer));
-    }
-  }
-  if (URL.get("type") === "CSV") {
-    var serverAnswer = getFromCSV(URL);
-    res.writeHead(200, headers);
-    res.end(JSON.stringify(serverAnswer));
-  }
-  if (URL.get("type") === "YAML") {
-    var serverAnswer = getFromYaml(URL);
-    res.writeHead(200, headers);
-    res.end(JSON.stringify(serverAnswer));
-  }
-}
-function watchPostUrl(req, res, headers) {
-  /*    функция, отвечающая за обработку get запросов, 
-  логика которой повязана на количестве типов, приехавших в объекте, 
-  если выбран не 1 тип вопроса - запишет во все, которые нужно */
-  jsonParser(req, res, (err) => {
-    if (err) console.log(err);
-    var URL = new URLSearchParams(req.url);
-    for (var i = 0; i < req.body.type.length; i++) {
-      if (req.body.type[i] === "JSON") {
-        var arrayFromJson = getFromJSONFile(URL, 1);
-        arrayFromJson.unshift(req.body);
-        fs.writeFileSync(
-          `questions/questions.json`,
-          JSON.stringify(arrayFromJson)
-        );
-      }
-      if (req.body.type[i] === "XML") {
-        var arrayFromXML = getFromXMLFile(URL, 1);
-        if (Array.isArray(arrayFromXML)) {
-          arrayFromXML.unshift(req.body);
-          fs.writeFileSync(
-            "questions/questions.xml",
-            convertToXML(arrayFromXML)
-          );
-        } else {
-          var newRecord = [];
-          if (arrayFromXML !== undefined) {
-            newRecord.push(arrayFromXML);
-          }
-          newRecord.unshift(req.body);
-          fs.writeFileSync("questions/questions.xml", convertToXML(newRecord));
-        }
-      }
-      if (req.body.type[i] === "YAML") {
-        var arrayFromYAML = getFromYaml(URL, 1);
-        arrayFromYAML.unshift(req.body);
-        fs.writeFileSync(
-          "questions/questions.yaml",
-          YAML.stringify(arrayFromYAML)
-        );
-      }
-      if (req.body.type[i] === "CSV") {
-        req.body.type[i] = ["CSV"];
-        var arrayFromCSV = getFromCSV(URL, 1);
-        arrayFromCSV.unshift(req.body);
-        fs.writeFileSync("questions/questions.csv", convertToCSV(arrayFromCSV));
-      }
-    }
-  });
-  res.writeHead(200, headers);
-  res.end("done");
-}
-function watchDeleteUrl(req, res, headers) {
-  /* Функция логики обработки delete запросов, основная на сравнении questions.date
-  При совпадении - объект будет удален из файла, файл перезапишется */
-  jsonParser(req, res, (err) => {
-    if (err) console.log(err);
-    var URL = new URLSearchParams(req.url);
-    if (URL.get("type") === "JSON") {
-      var arrayFromJson = getFromJSONFile(URL, 1);
-      for (var i = 0; i < arrayFromJson.length; i++) {
-        if (arrayFromJson[i].date === req.body.date) {
-          arrayFromJson.splice(i, 1);
-        }
-      }
-      fs.writeFileSync(
-        `questions/questions.json`,
-        JSON.stringify(arrayFromJson)
-      );
-    }
-    if (URL.get("type") === "XML") {
-      var arrayFromXML = getFromXMLFile(URL, 1);
-      if (!Array.isArray(arrayFromXML)) {
-        if (arrayFromXML.date === req.body.date) {
-          fs.writeFileSync(
-            `questions/questions.xml`,
-            "<questions></questions>"
-          );
-        }
-      } else {
-        for (var i = 0; i < arrayFromXML.length; i++) {
-          if (arrayFromXML[i].date === req.body.date) {
-            arrayFromXML.splice(i, 1);
-          }
-        }
-        fs.writeFileSync(`questions/questions.xml`, convertToXML(arrayFromXML));
-      }
-    }
-    if (URL.get("type") === "YAML") {
-      var arrayFromYAML = getFromYaml(URL, 1);
-      for (var i = 0; i < arrayFromYAML.length; i++) {
-        if (arrayFromYAML[i].date === req.body.date) {
-          arrayFromYAML.splice(i, 1);
-        }
-      }
-      fs.writeFileSync(
-        `questions/questions.yaml`,
-        YAML.stringify(arrayFromYAML)
-      );
-    }
-    if (URL.get("type") === "CSV") {
-      var arrayFromCSV = getFromCSV(URL, 1);
-      for (var i = 0; i < arrayFromCSV.length; i++) {
-        if (arrayFromCSV[i].date === req.body.date) {
-          arrayFromCSV.splice(i, 1);
-        }
-      }
-      fs.writeFileSync(`questions/questions.csv`, convertToCSV(arrayFromCSV));
-    }
-  });
-  res.writeHead(200, headers);
-  res.end("done");
-}
 
-function getFromJSONFile(URL, mode) {
-  /* Фунция чтения из файла JSON и фильтрации его, 
-  если нужно по темам, которые приходят в query параметрах */
-  var bufferFromJsonFile = fs.readFileSync(`questions/questions.json`);
-  var arrayFromJson = [];
-  var parsedBuffer = JSON.parse(bufferFromJsonFile);
-  if (URL.get("theme") === "ALLTHEMES" || mode === 1) {
-    return parsedBuffer;
-  } else {
-    for (var i = 0; i < parsedBuffer.length; i++) {
-      if (parsedBuffer[i].theme === URL.get("theme")) {
-        arrayFromJson.push(parsedBuffer[i]);
-      }
-    }
-  }
-  return arrayFromJson;
-}
 
-function getFromXMLFile(URL, mode) {
-  /* Фунция чтения из файла XML и фильтрации его, 
-  если нужно по темам, которые приходят в query параметрах */
-  var bufferFromXMLFile = fs.readFileSync(`questions/questions.xml`, "utf-8");
-  if (bufferFromXMLFile === "") {
-    fs.writeFileSync(
-      "questions/questions.xml",
-      "<questions><question></question></questions>"
-    );
-    bufferFromXMLFile = fs.readFileSync(`questions/questions.xml`, "utf-8");
-  }
-  var allFromBuffer = XMLparser.parse(bufferFromXMLFile);
-  var arrayFromXML = allFromBuffer.questions.question;
-  var themesArrayFromXML = [];
-  if (!Array.isArray(arrayFromXML)) {
-    var newArray = [];
-    newArray.push(arrayFromXML);
-    arrayFromXML = newArray;
-  }
-  if (URL.get("theme") === "ALLTHEMES" || mode === 1) {
-    return arrayFromXML;
-  } else {
-    for (var i = 0; i < arrayFromXML.length; i++) {
-      if (arrayFromXML[i].theme === URL.get("theme")) {
-        themesArrayFromXML.push(arrayFromXML[i]);
-      }
-    }
-  }
-  return themesArrayFromXML;
-}
-function convertToXML(array) {
-  var result = "<questions>";
 
-  for (var i = 0; i < array.length; i++) {
-    var wrapper = "<question>";
-    for (var key in array[i]) {
-      if (array[i] === null) {
-        continue;
-      }
-      wrapper += `<${key}>${array[i][key]}</${key}>`;
-    }
-    result += wrapper + "</question>";
-  }
 
-  result += "</questions>";
-  return result;
-}
-function convertToCSV(array) {
-  var result = "";
-  result += `${Object.keys(array[0])} \n`;
-  for (var i = 0; i < array.length; i++) {
-    if (array[i].theme === undefined) {
-      continue;
-    }
-    result += `${Object.values(array[i])} \n`;
-  }
-
-  return result;
-}
-function parseFromCSV(string) {
-  string = string.split("\n");
-  var result = [];
-  for (var i = 1; i < string.length; i++) {
-    var obj = {};
-    obj[string[0].split(",")[0]] = string[i].split(",")[0];
-    obj[string[0].split(",")[1]] = string[i].split(",")[1];
-    obj[string[0].split(",")[2]] = Number(string[i].split(",")[2]);
-    obj[string[0].split(",")[3]] = string[i].split(",")[3];
-    obj[string[0].split(",")[4]] = string[i].split(",")[4];
-    obj[string[0].split(",")[5].trim()] = Boolean(string[i].split(",")[5]);
-    result.push(obj);
-  }
-  return result;
-}
-function getFromCSV(URL, mode) {
-  var bufferFromCSVFile = fs.readFileSync("questions/questions.csv", "utf-8");
-  var arrayFromCSV = parseFromCSV(bufferFromCSVFile);
-  var themesCSVArray = [];
-  if (URL.get("theme") === "ALLTHEMES" || mode === 1) {
-    return arrayFromCSV;
-  } else {
-    for (var i = 0; i < arrayFromCSV.length; i++) {
-      if (arrayFromCSV[i].theme === URL.get("theme")) {
-        themesCSVArray.push(arrayFromCSV[i]);
-      }
-    }
-  }
-  return themesCSVArray;
-}
-
-function getFromYaml(URL, mode) {
-  var bufferFromYaml = fs.readFileSync("questions/questions.yaml", "utf-8");
-  var arrayFromYaml = yamlParser.load(bufferFromYaml);
-  var themesYAMLArray = [];
-  if (URL.get("theme") === "ALLTHEMES" || mode === 1) {
-    return arrayFromYaml;
-  } else {
-    for (var i = 0; i < arrayFromYaml.length; i++) {
-      if (arrayFromYaml[i].theme === URL.get("theme")) {
-        themesYAMLArray.push(arrayFromYaml[i]);
-      }
-    }
-    return themesYAMLArray;
-  }
-}
-
-//getFromYaml();
